#1680 my routes refactoring
