+++ app/Http/Controllers/Auth/Social/GithubController.php
@@ -67,7 +67,7 @@ class GithubController extends Controller
         Auth::login($user, true);
         flash()->success(__('auth.logged_in'));
 
-        return redirect()->route('my');
+        return redirect()->route('my.index');
     }
 
     private function sendFailedResponse($msg = null): RedirectResponse
