+++ resources/views/layouts/_nav.blade.php
@@ -61,7 +61,7 @@
                 <a class="dropdown-item" href="{{ route('settings.account.index') }}">{{ __('account.settings') }}</a>
               </li>
               <li>
-                <a class="dropdown-item" href="{{ route('my') }}">
+                <a class="dropdown-item" href="{{ route('my.index') }}">
                   {{ __('layout.nav.my_progress') }}
                 </a>
               </li>
